Rework CryptDataConverter example to use new PayloadCodec
